Safe server actions

+++ lib/data/logs.ts
@@ -4,10 +4,14 @@ import { logs, endpoints } from "../db/schema";
 import { eq, desc } from "drizzle-orm";
 import { revalidatePath } from "next/cache";
 import { db } from "../db";
-import { getErrorMessage } from "@/lib/helpers/error-message";
+import { authenticatedAction } from "./safe-action";
+import { z } from "zod";
 
 /**
  * Creates a log entry
+ *
+ * Helper function used in dynamic route for creating a log
+ * User does not need to be authenticated for this to happen
  */
 export async function createLog(
   type: "success" | "error",
@@ -29,43 +33,55 @@ export async function createLog(
 
 /**
  * Gets all logs for the user
+ *
+ * Protected by authenticatedAction wrapper
  */
-export async function getLogs(userId: string): Promise<LogRow[]> {
-  const logsData = await db
-    .select()
-    .from(logs)
-    .leftJoin(endpoints, eq(logs.endpointId, endpoints.id))
-    .where(eq(endpoints.userId, userId))
-    .orderBy(desc(logs.createdAt));
+export const getLogs = authenticatedAction.action(
+  async ({ ctx: { userId } }) => {
+    const logsData = await db
+      .select()
+      .from(logs)
+      .leftJoin(endpoints, eq(logs.endpointId, endpoints.id))
+      .where(eq(endpoints.userId, userId))
+      .orderBy(desc(logs.createdAt));
 
-  const data: LogRow[] = logsData.map((log) => ({
-    id: log.log.id,
-    type: log.log.type,
-    postType: log.log.postType,
-    message: log.log.message,
-    createdAt: log.log.createdAt,
-    endpointId: log.endpoint?.id || "",
-    endpoint: log.endpoint?.name || "",
-  }));
+    const data: LogRow[] = logsData.map((log) => ({
+      id: log.log.id,
+      type: log.log.type,
+      postType: log.log.postType,
+      message: log.log.message,
+      createdAt: log.log.createdAt,
+      endpointId: log.endpoint?.id || "",
+      endpoint: log.endpoint?.name || "",
+    }));
 
-  return data;
-}
+    return data;
+  }
+);
 
 /**
  * Deletes a log
+ *
+ * Protected by authenticatedAction wrapper
  */
-export async function deleteLog(id: string): Promise<
-  | {
-      error: string;
+export const deleteLog = authenticatedAction
+  .schema(z.object({ id: z.string() }))
+  .action(async ({ parsedInput: { id }, ctx: { userId } }) => {
+    const logWithEndpoint = await db
+      .select({
+        endpointUserId: endpoints.userId,
+      })
+      .from(logs)
+      .innerJoin(endpoints, eq(logs.endpointId, endpoints.id))
+      .where(eq(logs.id, id));
+
+    if (
+      !logWithEndpoint.length ||
+      logWithEndpoint[0].endpointUserId !== userId
+    ) {
+      throw new Error("You are not authorized for this action.");
     }
-  | undefined
-> {
-  try {
+
     await db.delete(logs).where(eq(logs.id, id));
     revalidatePath("/logs");
-  } catch (error: unknown) {
-    return {
-      error: getErrorMessage(error),
-    };
-  }
-}
+  });
